+++ src/test/kotlin/sigma/syntax/type_expressions/ArrayTypeConstructorTermTests.kt
@@ -1,35 +1,30 @@
 package sigma.syntax.type_expressions
 
-import sigma.semantics.BuiltinTypeScope
 import sigma.syntax.typeExpressions.TypeExpressionTerm
 import sigma.syntax.typeExpressions.TypeReferenceTerm
 import sigma.syntax.SourceLocation
-import sigma.syntax.typeExpressions.DictTypeTerm
+import sigma.syntax.typeExpressions.ArrayTypeConstructorTerm
+import sigma.semantics.types.ArrayType
 import sigma.semantics.types.BoolType
-import sigma.semantics.types.DictType
-import sigma.semantics.types.IntCollectiveType
+import sigma.evaluation.values.FixedTypeScope
 import sigma.evaluation.values.Symbol
 import kotlin.test.Test
 import kotlin.test.assertEquals
 
-class DictTypeDepictionTests {
+class ArrayTypeConstructorTermTests {
     object ParsingTests {
         @Test
         fun test() {
             val expression = TypeExpressionTerm.parse(
-                source = "{[K]: V}",
+                source = "[A*]",
             )
 
             assertEquals(
-                expected = DictTypeTerm(
+                expected = ArrayTypeConstructorTerm(
                     location = SourceLocation(lineIndex = 1, columnIndex = 0),
-                    keyType = TypeReferenceTerm(
-                        location = SourceLocation(lineIndex = 1, columnIndex = 2),
-                        referee = Symbol.of("K"),
-                    ),
-                    valueType = TypeReferenceTerm(
-                        location = SourceLocation(lineIndex = 1, columnIndex = 6),
-                        referee = Symbol.of("V"),
+                    elementType = TypeReferenceTerm(
+                        location = SourceLocation(lineIndex = 1, columnIndex = 1),
+                        referee = Symbol.of("A"),
                     ),
                 ),
                 actual = expression,
@@ -41,15 +36,18 @@ class DictTypeDepictionTests {
         @Test
         fun test() {
             val type = TypeExpressionTerm.parse(
-                source = "{[Int]: Bool}",
+                source = "[A*]",
             ).evaluate(
-                typeScope = BuiltinTypeScope,
+                typeScope = FixedTypeScope(
+                    entries = mapOf(
+                        Symbol.of("A") to BoolType,
+                    ),
+                ),
             )
 
             assertEquals(
-                expected = DictType(
-                    keyType = IntCollectiveType,
-                    valueType = BoolType,
+                expected = ArrayType(
+                    elementType = BoolType,
                 ),
                 actual = type,
             )
