+++ src/test/kotlin/sigma/syntax/type_expressions/DictTypeTermTests.kt
@@ -1,30 +1,35 @@
 package sigma.syntax.type_expressions
 
+import sigma.semantics.BuiltinTypeScope
 import sigma.syntax.typeExpressions.TypeExpressionTerm
 import sigma.syntax.typeExpressions.TypeReferenceTerm
 import sigma.syntax.SourceLocation
-import sigma.syntax.typeExpressions.ArrayTypeConstructorTerm
-import sigma.semantics.types.ArrayType
+import sigma.syntax.typeExpressions.DictTypeTerm
 import sigma.semantics.types.BoolType
-import sigma.evaluation.values.FixedTypeScope
+import sigma.semantics.types.DictType
+import sigma.semantics.types.IntCollectiveType
 import sigma.evaluation.values.Symbol
 import kotlin.test.Test
 import kotlin.test.assertEquals
 
-class ArrayTypeConstructorTests {
+class DictTypeTermTests {
     object ParsingTests {
         @Test
         fun test() {
             val expression = TypeExpressionTerm.parse(
-                source = "[A*]",
+                source = "{[K]: V}",
             )
 
             assertEquals(
-                expected = ArrayTypeConstructorTerm(
+                expected = DictTypeTerm(
                     location = SourceLocation(lineIndex = 1, columnIndex = 0),
-                    elementType = TypeReferenceTerm(
-                        location = SourceLocation(lineIndex = 1, columnIndex = 1),
-                        referee = Symbol.of("A"),
+                    keyType = TypeReferenceTerm(
+                        location = SourceLocation(lineIndex = 1, columnIndex = 2),
+                        referee = Symbol.of("K"),
+                    ),
+                    valueType = TypeReferenceTerm(
+                        location = SourceLocation(lineIndex = 1, columnIndex = 6),
+                        referee = Symbol.of("V"),
                     ),
                 ),
                 actual = expression,
@@ -36,18 +41,15 @@ class ArrayTypeConstructorTests {
         @Test
         fun test() {
             val type = TypeExpressionTerm.parse(
-                source = "[A*]",
+                source = "{[Int]: Bool}",
             ).evaluate(
-                typeScope = FixedTypeScope(
-                    entries = mapOf(
-                        Symbol.of("A") to BoolType,
-                    ),
-                ),
+                typeScope = BuiltinTypeScope,
             )
 
             assertEquals(
-                expected = ArrayType(
-                    elementType = BoolType,
+                expected = DictType(
+                    keyType = IntCollectiveType,
+                    valueType = BoolType,
                 ),
                 actual = type,
             )
